requirements.txt
pyaml==21.10.1
scipy==1.8.1
numpy==1.22.4
pandas==1.4.2
tqdm==4.64.0
lz4==4.0.1
roslibpy==1.3.0
bagpy==0.4.10
numba==0.52.0
nbconvert==6.5.0
matplotlib==3.5.3
torch==1.10.1+cu102
torchvision==0.11.2+cu102
torchaudio==0.10.1
gpytorch==1.6.0
tensorflow-cpu==2.9.0
